requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
gunicorn==20.1.0
charset-normalizer==3.1.0
cryptography==39.0.2
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.2
whitenoise==6.4.0
django-allauth==0.44.0
django-appconf==1.0.5
django-compressor==4.3.1
django-environ==0.4.5
django-extensions==3.2.1.dev0
django-google-maps==0.13.0
django-import-export==3.2.0
django-paypal==2.0
et-xmlfile==1.1.0
idna==3.4
logger==1.4
MarkupPy==1.14
MarkupSafe==2.1.2
mysqlclient==2.1.1
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.2
paypalrestsdk==1.13.1
Pillow==9.5.0
pycparser==2.21
PyJWT==2.6.0
pyOpenSSL==23.0.0
python-decouple==3.8
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
rcssmin==1.1.1
requests==2.28.2
requests-oauthlib==1.3.1
rjsmin==1.2.1
six==1.16.0
sqlparse==0.4.3
stripe==2.58.0
tablib==3.4.0
tzdata==2023.3
urllib3==1.26.15
Werkzeug==2.2.2
xlrd==2.0.1
xlwt==1.3.0
xmltodict==0.13.0
